Moves code which is only used in tests
